+++ pcs/cli/cluster_property/output.py
@@ -15,10 +15,12 @@ from pcs.common.str_tools import (
     format_list,
     format_name_value_default_list,
     format_name_value_list,
-    format_optional,
     indent,
 )
-from pcs.common.types import StringSequence
+from pcs.common.types import (
+    StringCollection,
+    StringSequence,
+)
 
 
 class PropertyConfigurationFacade:
@@ -26,9 +28,16 @@ class PropertyConfigurationFacade:
         self,
         properties: Sequence[CibNvsetDto],
         properties_metadata: Sequence[ResourceAgentParameterDto],
+        readonly_properties: StringCollection,
     ) -> None:
         self._properties = properties
         self._properties_metadata = properties_metadata
+        self._readonly_properties = readonly_properties
+        self._defaults_map = {
+            metadata.name: metadata.default
+            for metadata in self._properties_metadata
+            if metadata.default is not None
+        }
         self._name_nvpair_dto_map = (
             {
                 nvpair_dto.name: nvpair_dto
@@ -37,11 +46,6 @@ class PropertyConfigurationFacade:
             if self._properties
             else {}
         )
-        self._defaults_map = {
-            parameter.name: parameter.default
-            for parameter in self._properties_metadata
-            if parameter.default is not None
-        }
 
     @classmethod
     def from_properties_dtos(
@@ -50,8 +54,29 @@ class PropertyConfigurationFacade:
         properties_metadata_dto: ClusterPropertyMetadataDto,
     ) -> "PropertyConfigurationFacade":
         return cls(
-            properties_dto.sets,
+            properties_dto.nvsets,
             properties_metadata_dto.properties_metadata,
+            properties_metadata_dto.readonly_properties,
+        )
+
+    @classmethod
+    def from_properties_config(
+        cls, properties: ListCibNvsetDto
+    ) -> "PropertyConfigurationFacade":
+        return cls(
+            properties.nvsets,
+            [],
+            [],
+        )
+
+    @classmethod
+    def from_properties_metadata(
+        cls, properties_metadata_dto: ClusterPropertyMetadataDto
+    ) -> "PropertyConfigurationFacade":
+        return cls(
+            [],
+            properties_metadata_dto.properties_metadata,
+            properties_metadata_dto.readonly_properties,
         )
 
     @property
@@ -59,16 +84,63 @@ class PropertyConfigurationFacade:
         return self._properties
 
     @property
-    def properties_metadata(self) -> Sequence[ResourceAgentParameterDto]:
-        return self._properties_metadata
+    def readonly_properties(self) -> StringCollection:
+        return self._readonly_properties
 
-    @property
-    def defaults(self) -> dict[str, str]:
-        return self._defaults_map
+    def get_property_value(
+        self, property_name: str, custom_default=None
+    ) -> Optional[str]:
+        nvpair = self._name_nvpair_dto_map.get(property_name)
+        return nvpair.value if nvpair else custom_default
 
-    def get_property_value(self, property_name: str) -> Optional[str]:
-        property_dto = self._name_nvpair_dto_map.get(property_name)
-        return property_dto.value if property_dto else None
+    def get_property_value_or_default(
+        self, property_name: str, custom_default=None
+    ) -> Optional[str]:
+        value = self.get_property_value(property_name)
+        if value is not None:
+            return value
+        return self._defaults_map.get(property_name, custom_default)
+
+    @staticmethod
+    def _filter_names_advanced(
+        metadata: ResourceAgentParameterDto,
+        property_names: Optional[StringSequence] = None,
+        advanced: Optional[bool] = None,
+    ) -> bool:
+        return bool(
+            (
+                not property_names
+                and (advanced is None or metadata.advanced == advanced)
+            )
+            or (
+                property_names
+                and metadata.name in property_names
+                and (advanced is None or metadata.advanced == advanced)
+            )
+        )
+
+    def get_defaults(
+        self,
+        property_names: Optional[StringSequence] = None,
+        advanced: Optional[bool] = None,
+    ) -> dict[str, str]:
+        return {
+            metadata.name: metadata.default
+            for metadata in self._properties_metadata
+            if metadata.default is not None
+            and self._filter_names_advanced(metadata, property_names, advanced)
+        }
+
+    def get_properties_metadata(
+        self,
+        property_names: Optional[StringSequence] = None,
+        advanced: Optional[bool] = None,
+    ) -> Sequence[ResourceAgentParameterDto]:
+        return [
+            metadata
+            for metadata in self._properties_metadata
+            if self._filter_names_advanced(metadata, property_names, advanced)
+        ]
 
     def get_name_value_default_list(self) -> list[tuple[str, str, bool]]:
         name_value_default_list = [
@@ -99,20 +171,19 @@ def properties_to_text(
 
 def properties_to_text_with_default_mark(
     properties_facade: PropertyConfigurationFacade,
-    property_names_list: Optional[StringSequence] = None,
+    property_names: Optional[StringSequence] = None,
 ) -> list[str]:
     lines: list[str] = []
-    _id = (
-        properties_facade.properties[0].id
+    id_part = (
+        f" {properties_facade.properties[0].id}"
         if properties_facade.properties
         else ""
     )
-    id_part = format_optional(_id, template=" {}")
     lines = [f"Cluster Properties:{id_part}"]
     tuple_list = [
         item
         for item in properties_facade.get_name_value_default_list()
-        if not property_names_list or item[0] in property_names_list
+        if not property_names or item[0] in property_names
     ]
     lines.extend(indent(format_name_value_default_list(sorted(tuple_list))))
     return lines
@@ -125,23 +196,21 @@ def properties_to_cmd(
         options = [
             quote("=".join([nvpair.name, nvpair.value]))
             for nvpair in properties_facade.properties[0].nvpairs
+            if nvpair.name not in properties_facade.readonly_properties
         ]
-        return ["pcs property set"] + indent(options)
+        if options:
+            return ["pcs property set --force --"] + indent(options)
     return []
 
 
-def properties_defaults_to_text(
-    properties_facade: PropertyConfigurationFacade,
-    property_names: Optional[StringSequence] = None,
-) -> list[str]:
+def properties_defaults_to_text(property_dict: dict[str, str]) -> list[str]:
+    """
+    Convert property default values to lines of text.
+
+    property_dict -- name to default value map
+    """
     return format_name_value_list(
-        sorted(
-            [
-                (key, value)
-                for key, value in properties_facade.defaults.items()
-                if not property_names or key in property_names
-            ]
-        )
+        sorted((key, value) for key, value in property_dict.items())
     )
 
 
@@ -166,7 +235,12 @@ def _parameter_metadata_to_text(
     text.append(type_or_allowed_values)
     if metadata.default:
         text.append(f"Default: {metadata.default}")
-    return [metadata.name] + indent(text)
+
+    return [
+        f"{metadata.name} (advanced use only)"
+        if metadata.advanced
+        else metadata.name
+    ] + indent(text)
 
 
 def cluster_property_metadata_to_text(