+++ pcs/common/pacemaker/cluster_property.py
@@ -3,8 +3,10 @@ from typing import Sequence
 
 from pcs.common.interface.dto import DataTransferObject
 from pcs.common.resource_agent.dto import ResourceAgentParameterDto
+from pcs.common.types import StringCollection
 
 
 @dataclass(frozen=True)
 class ClusterPropertyMetadataDto(DataTransferObject):
     properties_metadata: Sequence[ResourceAgentParameterDto]
+    readonly_properties: StringCollection